simplify keyboard support
